Add files via upload
Based on JSON created from the website, have categorized three types of data: cells, grouped cells and pointers. 
Treated cells and grouped cells as objects with data or output. 
Treated pointers as linking "source" input and "target" output, and as indicator of functions to use (eg. here, "summary" statistics and "regression") 
Created a loop to run all the pointer functions.